% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfplot.R
\name{kfplot}
\alias{kfplot}
\title{Comparison Plot Method for List of OPSR kfold Objects}
\usage{
kfplot(
  x,
  i = c("ll_mean", "ll_p_mean", "r2"),
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{a list of objects of class \code{"opsr.kfold"}.}

\item{i}{the loss to extract (see 'Value' in \code{\link{loss}}). One of \code{"ll_mean"},
\code{"ll_p_mean"} or \code{"r2"}.}

\item{main}{a main title for the plot, see also \code{\link{title}}.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{...}{further arguments passed to \code{\link{boxplot}}.}
}
\value{
See 'Value' in \code{\link{boxplot}}.
}
\description{
Comparison Plot Method for List of OPSR kfold Objects
}
\seealso{
\code{\link{loss}}
}
