% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss}
\alias{loss}
\title{Out-Of-Sample Goodness of Fit Indicators}
\usage{
loss(object, data, test_ind, ...)
}
\arguments{
\item{object}{an object of class \code{"opsr"}.}

\item{data}{a data frame containing the variables in the model (usually from
a call to \code{\link{model.frame.opsr}}).}

\item{test_ind}{indicator for the test data.}

\item{...}{addditional arguments passed to \code{\link{opsr_from_opsr}}.}
}
\value{
A list containing the losses:
\itemize{
\item \code{z}: The regime/treatment membership of the test data.
\item \code{ll}: The out-of-sample log-likelihood of the test data.
\item \code{ll_mean}: \code{ll}averaged over each regime and in total.
\item \code{ll_p}: The out-of-sample log-likelihood of the test data for the selection
process.
\item \code{ll_p_mean}: \code{ll_p} averaged over each regime and in total.
\item \code{r2}: Coefficient of determination for each regime and in total.
}
}
\description{
Computes a loss for out-of-sample data points.
}
