% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsr_from_opsr.R
\name{opsr_from_opsr}
\alias{opsr_from_opsr}
\title{Refit an OPSR Model on New Data}
\usage{
opsr_from_opsr(object, data, ...)
}
\arguments{
\item{object}{an object of class \code{"opsr"}.}

\item{data}{a data frame containing the variables in the model.}

\item{...}{additional arguments passed to \code{\link{opsr}}.}
}
\value{
An object of class \code{"opsr"}
}
\description{
Refit an OPSR Model on New Data
}
