% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsr_kfold.R
\name{opsr_kfold}
\alias{opsr_kfold}
\title{k-Fold Cross-Validation for OPSR Model Fits}
\usage{
opsr_kfold(object, k = 10, verbose = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{"opsr"}.}

\item{k}{number of folds.}

\item{verbose}{if \code{TRUE}, prints working information during iterations over
folds.}

\item{...}{additional arguments passed to \code{\link{opsr_from_opsr}}}
}
\value{
An object of class \code{"opsr.kfold"}. See 'Details' section for more
information.
}
\description{
Computes out-of-sample losses.
}
\details{
The returned object is of length \code{k}, where each element contains the losses
as computed by \code{\link{loss}}.
}
\examples{
sim_dat <- OPSR::opsr_simulate()
dat <- sim_dat$data
dat$xo3 <- runif(n = nrow(dat))
dat$xo4 <- factor(sample(c("level1", "leve2", "level3"), nrow(dat), replace = TRUE))
f <- ys | yo ~ xs1 + xs2 + log(xo3) | xo1 + xo2 + xo3 + xo4 | xo1 + xo2 + xo3 | xo1 + xo2
fit <- OPSR::opsr(f, dat, printLevel = 0)
fit_select <- opsr_select(fit, loss = "bic", printLevel = 0)
kfold <- opsr_kfold(fit, printLevel = 0)
kfold_select <- opsr_kfold(fit_select, printLevel = 0)

## extract
ll_mean <- kfold["ll_mean"]
list2df(ll_mean)

## plot
colvec <- c("#ff8811","#48a9a6")
kfplot(list(kfold, kfold_select), col = colvec)
legend("bottomleft", legend = c("fit", "fit_select"), fill = colvec,
       title = "Model", title.font = 2, bty = "n")
}
\seealso{
\code{\link{opsr_from_opsr}}, \code{\link{loss}}, [\verb{[.opsr.kfold}]
}
