% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsr_step.R
\name{opsr_step}
\alias{opsr_step}
\title{Step Function for OPSR Model Fits}
\usage{
opsr_step(object, pval, log = new.env(), .step = 1, ...)
}
\arguments{
\item{object}{an object of class \code{"opsr"}.}

\item{pval}{coefficients with p-values < \code{pval} are dropped.}

\item{log}{environment to keep track of changes to \code{object} (in particular
variables being eliminated).}

\item{.step}{used to generate identifier in \code{log} environment. Used in
\code{\link{opsr_select}}.}

\item{...}{additional arguments passed to \code{\link{update}} (and hence \code{\link{opsr}}).}
}
\value{
An object of class \code{"opsr"}.
}
\description{
Excludes all coefficients with p-values below \code{pval} and fits again.
}
\seealso{
\code{\link{opsr_select}}
}
