% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{package}
\name{OPSRtools-package}
\alias{OPSRtools}
\alias{OPSRtools-package}
\title{OPSRtools: Tools for OPSR}
\description{
OPSR estimates ordered probit switching regression models - a Heckman type endogenous switching regression model with an ordinal selection/treatment and continuous outcomes. OPSRtools contains estimation and post-estimation routines such as step-wise model selection, k-fold cross-validation, treatment effect computations and some visualizations.
}
\author{
\strong{Maintainer}: Daniel Heimgartner \email{d.heimgartners@gmail.com} (\href{https://orcid.org/0000-0002-0643-8690}{ORCID}) [copyright holder]

}
