% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_opsr_kfold.R
\name{plot.opsr.kfold}
\alias{plot.opsr.kfold}
\title{Plot Method for OPSR kfold Objects}
\usage{
\method{plot}{opsr.kfold}(
  x,
  i = c("ll_mean", "ll_p_mean", "r2"),
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"opsr.kfold"}.}

\item{i}{the loss to extract (see 'Value' in \code{\link{loss}}). One of \code{"ll_mean"},
\code{"ll_p_mean"} or \code{"r2"}.}

\item{main}{a main title for the plot, see also \code{\link{title}}.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{...}{further arguments passed to \code{\link{boxplot}}.}
}
\value{
See 'Value' in \code{\link{boxplot}}.
}
\description{
Plot Method for OPSR kfold Objects
}
\seealso{
\code{\link{loss}}
}
