% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_opsr_ate.R
\name{print.opsr.ate}
\alias{print.opsr.ate}
\title{Print Method for OPSR ATE Objects}
\usage{
\method{print}{opsr.ate}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"opsr.ate"}.}

\item{...}{further arguments passed to \code{\link{summary.opsr.ate}}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print Method for OPSR ATE Objects
}
\details{
This is just a wrapper around \code{\link{summary.opsr.ate}} and a subsequent call to
\code{\link{print.summary.opsr.ate}}.
}
\seealso{
\code{\link{print.summary.opsr.ate}}
}
