% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_opsr_select.R
\name{print.opsr.select}
\alias{print.opsr.select}
\title{Print Method for OPSR Select Objects}
\usage{
\method{print}{opsr.select}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  print.call = TRUE,
  print.elim.hist = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"opsr.select"}.}

\item{digits}{minimum number of significant digits to be used for most numbers (passed to \code{\link[stats:printCoefmat]{stats::printCoefmat}}).}

\item{print.call}{if \code{TRUE}, prints the underlying \code{\link{opsr_select}} call.}

\item{print.elim.hist}{if \code{TRUE}, prints the elimination history. See
'Details' section for more information.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints \code{x} in 'pretty' form and returns it invisibly.
}
\description{
Prints the original model and the final model (winner) as well as the
stepwise model path.
}
