% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_opsr_ate.R
\name{print.summary.opsr.ate}
\alias{print.summary.opsr.ate}
\title{Print Method for Summary OPSR ATE Objects}
\usage{
\method{print}{summary.opsr.ate}(x, digits = max(3L, getOption("digits") - 3L), print.call = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{"summary.opsr.ate"}.}

\item{digits}{minimum number of significant digits to be used for most numbers (passed to \code{\link[stats:printCoefmat]{stats::printCoefmat}}).}

\item{print.call}{if \code{TRUE}, prints the underlying call.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints \code{x} in 'pretty' form and returns it invisibly.
}
\description{
Print Method for Summary OPSR ATE Objects
}
