% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_te.R
\name{print.te}
\alias{print.te}
\title{Print Method for TE Objects}
\usage{
\method{print}{te}(x, digits = max(3L, getOption("digits") - 3L), signif.legend = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{"te"}.}

\item{digits}{minimum number of significant digits to be used for most numbers
(passed to \code{\link[stats:printCoefmat]{stats::printCoefmat}}).}

\item{signif.legend}{if \code{TRUE}, a legend for the 'significance stars' is printed.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints \code{x} in 'pretty' form and returns reformatted treatment effects
invisibly.
}
\description{
Print Method for TE Objects
}
