% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[.opsr.kfold}
\alias{[.opsr.kfold}
\title{Extract Method for OPSR kfold Object}
\usage{
\method{[}{opsr.kfold}(x, i)
}
\arguments{
\item{x}{an object of class \code{"opsr.kfold"}.}

\item{i}{the loss to extract (see 'Value' in \code{\link{loss}}).}
}
\value{
A list of length k (where k is the number of folds).
}
\description{
Extract Method for OPSR kfold Object
}
\seealso{
\code{\link{loss}}
}
