% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{timeuse}
\alias{timeuse}
\title{TimeUse+ Data}
\format{
Data frame with numeric and factor columns
\describe{
\item{uid}{Respondent ID}
\item{weight}{Sample weight}
\item{wfh}{Telework status}
\item{wfh_days}{Telework frequency (days/week)}
\item{weekly_km}{Weekly kilometers traveled}
\item{log_weekly_km}{Log weekly kilometers traveled}
\item{weekly_n}{Weekly number of trips}
\item{log_weekly_n}{Log weekly number of trips}
\item{start_tracking}{Month when tracking started}
\item{commute}{One-way commute distance (km)}
\item{log_commute_km}{Log one-way commute distance}
\item{age}{Age}
\item{car_access}{Car access}
\item{dogs}{Has dogs}
\item{driverlicense}{Has driverlicense}
\item{educ_higher}{Higher education}
\item{fixed_workplace}{Has fixed workplace}
\item{grocery_shopper}{Main responsible for grocery shopping}
\item{hh_income}{Household income}
\item{hh_size}{Household size}
\item{isco_clerical}{ISCO clerical}
\item{isco_craft}{ISCO craft}
\item{isco_elementary}{ISCO elementary}
\item{isco_managers}{ISCO managers}
\item{isco_plant}{ISCO plant and machine}
\item{isco_professionals}{ISCO professionals}
\item{isco_service}{ISCO service and sales}
\item{isco_agri}{ISCO agriculture}
\item{isco_tech}{ISCO technicians}
\item{married}{Married}
\item{n_children}{Number of children}
\item{freq_onl_order}{Frequent online shopper}
\item{parking_home}{Parking available at home}
\item{parking_work}{Parking available at work}
\item{permanent_employed}{Permanent employment}
\item{rents_home}{Tenant}
\item{res_loc}{Residential location}
\item{sex_male}{Male}
\item{shift_work}{Works in shifts}
\item{swiss}{Nationality Swiss}
\item{vacation}{Vacation during study}
\item{workload}{Workload (in \% of full-time equivalent)}
\item{young_kids}{Has young kids}
\item{wfh_allowed}{Employer allows telework}
\item{teleworkability}{Number of days worked from home during COVID-related lockdowns}
}
}
\usage{
timeuse
}
\description{
Travel diary data from the TimeUse+ study.
}
\references{
Winkler C, Meister A, Axhausen KW (2024). “The TimeUse+ Data Set: 4 Weeks of Time
Use and Expenditure Data Based on GPS Tracks.” Transportation, pp. 1–27. doi:10.1007/s11116-024-10517-1.
}
\keyword{datasets}
