% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsr_ate.R
\name{opsr_ate}
\alias{opsr_ate}
\title{Treatment Effect Computations for OPSR Model Fits}
\usage{
opsr_ate(object, type, weights = NULL, ...)
}
\arguments{
\item{object}{object an object of class \code{"opsr"}.}

\item{type}{see \code{\link{predict.opsr}} for details.}

\item{weights}{a vector of weights. If \code{NULL} then weights from \code{object} will
be used.}

\item{...}{additional arguments passed to \code{\link{predict.opsr}}.}
}
\value{
An object of class \code{"opsr.ate"}.
}
\description{
Treatment Effect Computations for OPSR Model Fits
}
\details{
This function only prepares the input to a further call to \code{\link{summary.opsr.ate}}.
}
\seealso{
\code{\link{summary.opsr.ate}}
}
