% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsr_kfold.R
\name{opsr_kfold}
\alias{opsr_kfold}
\title{k-Fold Cross-Validation for OPSR Model Fits}
\usage{
opsr_kfold(object, k = 10, verbose = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{"opsr"}.}

\item{k}{number of folds.}

\item{verbose}{if \code{TRUE}, prints working information during iterations over
folds.}

\item{...}{additional arguments passed to \code{\link{opsr_from_opsr}}}
}
\value{
An object of class \code{"opsr.kfold"}. See 'Details' section for more
information.
}
\description{
Computes out-of-sample losses.
}
\details{
The returned object is of length \code{k}, where each element contains the losses
as computed by \code{\link{loss}}.
}
\seealso{
\code{\link{opsr_from_opsr}}, \code{\link{loss}}, [\verb{[.opsr.kfold}]
}
